Add initial content
This adds starting content for the Mu Developer Operations (DevOps)
repo.

This repo (at least at this time) is primarily composed of Azure
Pipelines YAML files used as common resources between many Project
Mu repositories.

Repos are intended to bind against a source repository at pipeline
build time to expose the templates used for the build.
https://docs.microsoft.com/en-us/azure/devops/pipelines/process/resources?view=azure-devops&tabs=schema#define-a-repositories-resource

Each "leaf" YAML file (the YAML file used in the actual pipeline) is
expected to define a repository resource onto this repo and include
the appropriate top-level template file.

Typically, this is:
  - Jobs/PrGate.yml
  - Steps/BuildPlatform.yml

Each template provides parameter customization for common config
points of steps implemented in the template.

This change is considered an initial refactoring and convergence of
build functionality with the goal to replicate pipeline behavior
using these templates in this converged repo. It will be iterated
and impoved upon in future changes.

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>