Steps/PrGate.yml: Add pkg_count condition to build steps
Adds a condition to build steps to only execute if pkg_count is
greater than zero. The pkg_count is set to a default value of 1.
Stuart PR Eval might set it to zero if that is used and no packages
need to be built. When it does so, it also clears the package name
variable. So not only is the a performance optimization, it is
required if Stuart PR Eval runs and determines no packages need
to be built.

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>