PrGate.yml: Add support for publishing code coverage results
Introduces a new parameter called "extra_build_args" that can be used
to specify additional flags to pass to build commands.

This is used to pass flags needed to generate code coverage results
in this change.

The coverage.xml and coverage.html files added to the log copy step
will not have an impact if code coverage is not enabled.

For now, the code coverage publication step depends on the platform
setting "CODE_COVERAGE=TRUE" in the "extra_build_args" parameter.

This can be refactored/generalized as needeed in the future.

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>