Update code to align with previous isOutput bug behavior
Since "isOutput" was spelled incorrectly ("isOutpout"), the resulting
behavior was equivalent to "isOutput" not being specified at all.

Subsequent code has been written such that it refers to the variable
as a variable set in the same job rather than a multi-job variable as
"isOutput" is used for - the code simply refers to the variable name
rather than <step>.<var name>.

This change simply removes "isOutput" so the original set of code
assumptions continue to work as before.

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>