Add reusable Label Sync GitHub workflow
Adds a new workflow and accompanying manifest that specifies labels
used in Project Mu to exclusively control label management from
this centralized repository.

Repos can be specified directly in the GitHub Action run in this repo
(.github/workflows/LabelSyncer.yml). The action to sync labels across
Project Mu repos is triggered on any push to the label manifest file
in the main branch of this (the Mu DevOps) repo.

This has the benefit of reducing developer time working with labels
but more importantly makes labels consistent across Project Mu repos
and tracks changes to labels in Mu DevOps source history.

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>