Enable Label Sync GitHub Action
Adds an action to automatically sync labels used in this repository
from a centralized file in Project Mu DevOps (this repo).

New labels should be added to the `Labels.yml` file in mu_devops.

Repo-specific labels are allowed to be defined. Those can either be
created in a repo-local config YAML file (preferred) or manually
in the "Labels" section of the repo.

Note: This actually enables labels to be synced in the Mu DevOps
repository using the common/reusable workflow in Mu DevOps.

- [ ] Breaking change?
  - Will this change break pre-existing builds or functionality without action being taken?
  **No**

On fork repositories.

Now downstream integration required, only impacts this repo

Repo users should understand label definitions and read instructions
noted in this change regarding how to modify labels in the future.

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>