label-issues: Invert functional impact label logic
The checkbox exposed to contributors in a PR to determine whether
the PR has a functional impact states:

"Impacts functionality?"

This uses positive logic to simplify understanding by the user.

Currently the `impact:non-functional` label is applied if the
checkbox is checked.

This change inverts the logic to apply the label as expected.

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>