.sync: Set recent workflows to a fixed version of mu_devops
When some workflows were added recently, the reusable workflow file
that the leaf files depend on in mu_devops did not exist yet.

Therefore, those leaf files depended on the `main` branch instead
of a mu_devops release tag as the files would be in the `main`
branch as soon as the PR was merged.

Now that all of the reusable workflows are in a tagged release
(as of the `v1.3.0` release), this change sets them to a fixed version
of Mu DevOps as well.

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>