Split approval and merge workflows
Splitting the workflows provides several benefits:

1. Simplifies logic due to jobs being attached to separate triggers
   relevant for the specific job.
2. Reduces number of status checks shown in PRs due to certain steps
   being skipped in irrelevant circumstances.
3. Allows better workflow reuse.

The reviewers are also dropped from dependabot PRs since the pending
team reviewer is (1) not acted upon (2) can delay PR merging if the
reviewer team has not reviewed the PR.

The version is updated in anticipation of this change. It will be
modified if the expected version changes before this PR is merged.

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>