release-draft-config.yml: Add breaking changes section
Breaking changes are identified with the GitHub label:
  `impact:breaking-change`

Prior to this change, that label rolled the major version of the
repo.

Now, the label also places corresponding changes into a "Breaking
Changes" section of the release notes so it is easy for consumers
to see breaking changes in a release.

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>