Add a dependabot label and exclude it from release notes
The `release-drafter` GitHub action has an `exclude-contributors`
configuration option that excludes usernames from the final
contributors. This does not appear to work for dependabot.

The `exclude-labels` configuration option does work. This change
adds a new label `type:dependabot` to dependabot PRs (via the
dependabot config file) and uses the `exclude-labels` option in
the `release-drafter` config file to exclude those PRs from release
notes.

Submodules are currently updated by either dependabot or the
`submodule-release-updater` GitHub action local to mu_devops. Since
submodules are recursive to repo consumers and to keep those
consistently in the release notes, the `type:dependabot` label is
not added to dependabot submodule updates keeping those updates in
the release notes.

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>