Adding support for self-hosted pools (#154)
This change added support for self-hosted agent pools for potentially
onboarding Windows on ARM systems as there is no plan for Microsoft
hosted Windows ARM agents as of today.

The main change is around the flag of installing tools or not. The
existing routine will expect the incoming flag being a boolean type,
which does not work when this parameter is being populated using a
variable (string). This change will update the flag to string type as
well.

After changing, the only file being synced to individual repos is
`MuDevOpsWrapper.yml`. And each repo can configure their own build
matrix for build target, architecture list, run on selfhost agents or
not, etc.

This is considered a breaking change, due to:
1. `Steps/PrGate.yml` updated the `extra_install_steps` to be invoked
before `stuart_setup` and `stuart_update`. But this change does not
knowingly break the existing consumers.
2. `Jobs/PrGate.yml` updated the parameters to accept
`parameters.build_matrix`, which needs the consumers to update top level
yaml script to comply with the updated template.

---------

Co-authored-by: Michael Kubacki <michael.kubacki@microsoft.com>