Publish coverage on even with skipped builds (#198)
Current pipeline job for publishing coverage will skip the jobs if there
are a few jobs being skipped. However, some skips are expected, such as
some jobs that should be run on selfhosted agents, but no agents are
available on the public site.

This change will update the publish coverage job's condition to ignore
the skipped conditions, such that it will always run the job as long as
there the previous job(s) not failing or being cancelled.