Create simple tag generator for ADO pipelines (#222)
Creates a script and pipeline that allows for automated tagging based on
release version and commit messages. This template should be called for
an Azure Devops hosted repository. The consumer of this template is
expected to ensure the proper permissions for the build agent to be able
to create the tag and commit tag notes.

The TagGenerator Script provides the following functionality:
- Scans git history for the most recent matching tag, e.g. 202302.5.10
- Generates the new tag version, incrementing the "major" version based
on the presence of breaking changes.
- Generates release notes including commits by type, links to Azure
Devops PRs, and contributors.