.sync/codeql: Add cargo make and install rust-src
Updates the workflow to be able to build Rust code if present in
a repo. The Rust toolchain is installed by default, so this simply
installs `cargo make` and installs `rust-src`.

`cargo-make` is cached so it can be reused across runs until the
next version is published in the upstream repo.

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>