.sync/Makefile.toml: Resolve cargo-tarpaulin breaking changes
Tarpaulin 0.27 was released on September 17, 2023:
https://github.com/xd009642/tarpaulin/releases/tag/0.27.0

The clap crate (Command Line Argument Parser) dependency within
tarpaulin was upgraded in the 0.27 release from an old version
(v2) to the latest major version (v4):

```
  Upgraded from clap v2 to v4. This has a few changes, notably any
  arguments which can be specified more than once require multiple
  entries so --run-types doc test needs to be turned into
  --run-types doc --run-types test
```

This means passing packages to a single `-p` parameter as a comma-
separated list is no longer supported:

```
  [cargo-make] Execute Command: "cargo" "tarpaulin" "-p" \
    "HelloWorldRustDxe,RustBootServicesAllocatorDxe"

  cargo_tarpaulin::config: Creating config
  cargo_tarpaulin: Running Tarpaulin
  cargo_tarpaulin: Building project
  cargo_tarpaulin::cargo: Cleaning project
  error: invalid character `,` in pkgid:
    `HelloWorldRustDxe,RustBootServicesAllocatorDxe`, characters must
    be Unicode XID characters (numbers, `-`, `_`, or most letters)
```

Providing users the ability to pass packages separated by commas to
`cargo make coverage` is convenient and assumed in our documentation
and wrapper scripts.

This change retains the same user-facing interface to `cargo make coverage`
while using a duckscript within the cargo makefile to transform the
list to the format accepted by tarpaulin.

The package split is also applied to other commands for consistency.

The `PACKAGE_TARGET` env variable is removed since `CARGO_MAKE_TASK_ARGS`
is handled elsewhere in the file now.

Duckscript is useful because it is readily embedded in cargo-make so
no additional dependencies are required and it is cross-platform.

---

In addition, a minor fix is made by changing `Html` to `html` for the
following issue:

```
  [cargo-make] Execute Command: "cargo" "tarpaulin" "--out" "Html"
    "--exclude-files" "**/tests/*" "--output-dir" "D:\\src\\mu_plus/target"
  error: invalid value 'Html' for '--out [<FMT>...]'
    [possible values: json, stdout, xml, html, lcov]

    tip: a similar value exists: 'html'
```

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>