Conditionalize RustSetupSteps.yml to Non-Self-Hosted Agents (#246)
The current rust setup script embedded for package CI build will cause
failures on selfhosted agent pipelines.

On Windows AARCH64 agents, the failure is due to the installation of
toolchain is currently specified to be x86_64, which will not succeed on
AARCH64 systems in the self-host pool. In addition, the update will
apply to installed cargo, which could cause conflict when multiple
agents run on the same host system.

On Linux agents, the failure is due to modifications made to container
permissions. Self-hosted Linux agents do not use containers.

The change here removes the toolchain installation steps for selfhost
agent matrix builds as it assumes these systems preset the needed
environment properly.