Add UploadCodeCoverage Template  (#258)
Add a new template, UploadCodeCoverage, that can be used to upload a
code coverage document (such as cobertura) to codecov.io.

Uses the `-Z` flag, which will cause the step to fail if uploading fails
for any reason.

See:
https://dev.azure.com/projectmu/mu/_build/results?buildId=55185&view=results
for a working example.

Closed #257 but this should be usable in non-rust environments. I will
start uploading all code coverage (such as for mu_basecore) once I have
`stuart_report` merged and being used to clean up the code coverage
data.