MuDevOpsWrapper: Allow coverage uploads to ado or codecov.io (#270)
## Description

This change reverts the changes made in #266 in favor of performing the
reformatting in the HostBasedUnitTestRunner plugin. From there it allows
the consumer of `MuDevOpsWrapper` to configure where code coverage data
is uploaded to, either `ado` (The azure devops pipeline that is running)
or `codecov` (The codecov account associated with the repository).

* Uploading to ado remains the same - A final job runs that merges
coverage data and uploads it to azure pipelines
* Due to the advanced capabilities, uploading to codecov runs per matrix
job and uploads using flags to separate code coverage data between
packages.

## Integration Instructions

The previous logic remains the default, that is to say code coverage is
uploaded to the azure devops pipeline that is running. Should a consumer
set `coverage_publish_target` to 'codecov', they must ensure that
`CODECOV_TOKEN` is set as a private variable in the pipeline.