.sync/workflows/codeql-platform: Update irrelevant plugin removal dir
Updates the directory used for irrelevant plugin removal in this
workflow to match codeql.yml which using the pytool directory rather
than the directory the CodeQL plugin is found in.

This is important because the CodeQL plugin moved from .pytool to
BaseTools in Mu release 202311 so the plugin dir needs to be fixed
rather than relative to the CodeQL plugin location.

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>