Prevent rustup self update (#327)
By default `rusup` performs a self-update when installing a new rustc
toolchain, Sometimes, the files are locked by windows, and the self
update can fail as described here:
https://github.com/rust-lang/rustup/issues/2441

Since this is a pipeline run, we have no need to update rustup, and can
rely on the version provided by the runner. By disabling the self
update, the file conflict `The process cannot access the file because it
is being used by another process` no longer appears.