.sync/Dockerfile: Add Python 3.12 installation steps (#383)
Some additional steps are needed to move the container build to Python
3.12.

The Deadsnakes PPA is needed to get Python 3.12. Add a step to add that
repo and then use apt to install Python 3.12 from it. gnupg is needed to
add the key for Deadsnakes. Distutils was removed from Python 3.12 but
we still need it to build QEMU so that is added as a separate
installation.

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>