Add release branch backport workflow (#386)
Closes #372 

This new workflow will attempt to automatically cherry-pick marked
contributions to a development branch to its corresponding release
branch. If a merge conflict occurs, the commit is committed to a new
branch with merge markers and then a PR is created into the target
branch with those markers. The PR is labeled with
`type:release-merge-conflict` to indicate that it needs manual
resolution.

The PR (if created) is expected to fail compilation and status checks of
course due to the merge conflict markers. A human should then checkout
the PR branch, resolve the conflicts, and push the changes back to the
PR branch.

---

- To mark a PR going into the development branch so that it should be
cherry-picked to the release branch, add the `type:backport` label to
the PR.

---

Notes:

- The workflow is synced to all repos that currently have a `dev`
branch.
- Each repo that `backport-to-release-branch.yml` is synced to must have
a `CHERRY_PICK_TOKEN` defined with repo write permission.

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>