Update Release Drafter for Dev Branches
Changes the way release drafter handles drafting releases on repos
with release branches (which have dev branches now).

Ultimately, a release is drafted for the dev branch and the release
branch. The releases are differentiated by both their release title
and tag.

- Release Branch
  - Title: `release-v<version>`
  - Tag: `v<version>`
- Dev Branch
  - Title: `dev-v<version>`
  - Tag: `dev-v<version>`

Note that the tag for the release branch follows the same convention
as existing release tags.

The "release branch" release includes all pull requests made to the
dev branch with the `type:backport` label since the last "dev branch"
release. For this reason, the "dev branch" and "release branch"
should be released at the same time. Then, this effectively results
in the "release branch" having all relevant changes since the last
release. The "dev branch" release will be based at the same point in
history as the "release branch" release but include all changes not
just those with the `type:backport` label.

The "release branch" release for the current release branch should be
marked as "latest". For example, if "release/202311" and
"release/202405" exist, the "release/202405" "release branch" release
would be marked as latest.

The `release-drafter/release-drafter` action is still used.

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>