DownloadCargoBinaryFromGitHub: Find most recent release with assets
Currently, the `cargo-binstall` most recent release in the binstall
repo does not actually have the binstall binary. This change allows
the search to find the most recent release that has the expected
binaries to handle the case when other releases might also be in
release list that do not have the expected binaries.

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>