codeql: Go back to pull_request (#464)
b9c5931 moved to pull_request_target to use the GitHub app to derive
auth tokens. This may not allow GitHub to have the propoer context for
the changes in the PR. This goes back to pull_request. Since only read
permission is needed for the token, this sets uses the default token to
make authenticated API calls.

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>