Rust setup: Use UID / GID (#484)
Depending on the pipeline runner, the account name created in the
container may be `azure_pipelines_docker` or `docker_azpcontainer`,
leading to issues in setting the permissions that allow us to create or
move files for rust setup / rust testing. It was seen that regardless of
the name, the actual UID / GID that they are using is always 1001, so we
instead hardcode the UID/GID value instead of the name.

This has proven to work here:


https://dev.azure.com/projectmu/mu/_build/results?buildId=96553&view=logs&j=a0dae8b7-2e4d-50ba-22df-26bfc97f6333&t=f5f7ed98-5e80-50ec-f2eb-a7de1997e1f1